Modified between-city movement/infection
Based off talks with Tariq and Kaare. The number of infected who move
FROM a city is drawn from a poisson distribution where the mean
(lambda) is = (number Infected* * (percentage of population that
travels). We then track where the Infected move to. We don’t track S
and R because we assume symmetric travel.